Change the type of display port from byte to int
This is necessary because we changed the return type of
DisplayAddress.getPort() in the framework.

Test: m
Bug: 153334857

+++ car-lib/src/android/car/media/CarAudioManager.java
@@ -540,7 +540,7 @@ public final class CarAudioManager extends CarManagerBase {
      * @hide
      */
     @RequiresPermission(Car.PERMISSION_CAR_CONTROL_AUDIO_SETTINGS)
-    public int getZoneIdForDisplayPortId(byte displayPortId) {
+    public int getZoneIdForDisplayPortId(int displayPortId) {
         try {
             return mService.getZoneIdForDisplayPortId(displayPortId);
         } catch (RemoteException e) {
